requirements.txt
# Django libraries
Django==4.0.5 
djangorestframework==3.13.1


# Third party libraries

